# SOHO-7331 - e2e CI preparation, and optimization