--- a/test/helpers/axe-page-objects.js
+++ b/test/helpers/axe-page-objects.js
@@ -4,10 +4,10 @@ const rules = require('./default-axe-options.js');
 const AxePageObject = async function (theme) {
   let res;
   if (theme === 'high-contrast') {
-    res = AxeBuilder(browser.driver).exclude('header').analyze();
+    res = await AxeBuilder(browser.driver).exclude('header').analyze();
   }
 
-  res = AxeBuilder(browser.driver).configure({ rules }).exclude('header').analyze();
+  res = await AxeBuilder(browser.driver).configure({ rules }).exclude('header').analyze();
 
   return res;
 };