--- a/test/helpers/e2e-utils.js
+++ b/test/helpers/e2e-utils.js
@@ -3,4 +3,9 @@ module.exports = {
   isFF: () => browser.browserName === 'firefox',
   isSafari: () => browser.browserName === 'safari',
   isChrome: () => browser.browserName === 'chrome',
+  setPage: async (url) => {
+    const pageurl = `${browser.baseUrl + url}?theme=${browser.params.theme}`;
+    await browser.waitForAngularEnabled(false);
+    await browser.driver.get(pageurl);
+  }
 };