> Explain the **details** for making this change. What existing problem does the pull request solve?

Tests are slow. Remove sleep command, and replace with wait command
 - Currently: time env ENTERPRISE_THEME='dark' npm run e2e:local  3.70s user 0.65s system 8% cpu **53.996** total
- Previously: time env ENTERPRISE_THEME='dark' npm run e2e:local  4.69s user 0.84s system 4% cpu **2:16.29** total

Clearly label Protractor configurations
Skip failing WCAG tests

<!-- Example: When "Adding a function to do X", explain why it is necessary to have a way to do X. -->

> **Related issue (required)**:

https://jira.infor.com/browse/SOHO-7331
<!-- Provide a link to the related issue to this Pull Request. ***Please do not open a PR without a related issue.*** -->

> **Steps necessary to review your pull request (required)**:

Debug like so... Maybe master url or some other build url...
```sh
npm start && env ENTERPRISE_THEME='high-contrast' npm run e2e:local -- --baseUrl='http://master-enterprise.demo.design.infor.com'
```
Isolate a test, and run `npm run e2e:bs`
<!-- What does someone need to do to confirm that this PR is ready to merge? Please include the exact commands you ran and their output, screenshots / videos if the pull request changes UI. You can skip this if you're fixing a typo or the change is very obvious in the diff. -->

_WIP_
Maybe run on the nightly, every few hours,`npm run e2e:ci`

> Other Details:

<!-- Please add `closes #XXXX` in your comment to auto-close the issue that your PR fixes (if exists). -->

<!-- After submitting your PR, please check back to make sure tests pass on Travis. -->