@clepore 2018-05-23
Manual merge per Travis being down